// src/frontend/src/backend.d.ts
export interface LineItem {
    rate: bigint;
    itemName: string;
    discount: bigint;
    quantity: bigint;
    taxRate: bigint;
    taxType: string;
}